map-reduce-func/master-code/requirements.txt
functions-framework==3.*
google-cloud-storage==2.13.0
six==1.16.0
requests==2.26.0